Fix remaining int64_t/intptr_t/SQLLEN type issues with OSX

+++ cpp/cpp_odbc/Library/cpp_odbc/multi_value_buffer.h
@@ -7,12 +7,12 @@ namespace cpp_odbc {
 
 struct writable_buffer_element {
 	char * data_pointer;
-	int64_t & indicator;
+	intptr_t & indicator;
 };
 
 struct buffer_element {
 	char const * data_pointer;
-	int64_t const & indicator;
+	intptr_t const & indicator;
 };
 
 /**
@@ -69,7 +69,7 @@ public:
 	 *        StrLen_or_IndPtrs appear in ODBC calls
 	 * @return A pointer to the length/indicator array
 	 */
-	int64_t * indicator_pointer();
+	intptr_t * indicator_pointer();
 
 	/**
 	 * @brief Provides read-only raw access to a contiguous strip of memory which
@@ -77,7 +77,7 @@ public:
 	 *        is NULL).
 	 * @return A pointer to the length/indicator array
 	 */
-	int64_t const * indicator_pointer() const;
+	intptr_t const * indicator_pointer() const;
 
 	/**
 	 * @brief Read/write access to the element with index element_index
@@ -96,7 +96,7 @@ public:
 private:
 	std::size_t element_size_;
 	std::vector<char> data_;
-	std::vector<int64_t> indicators_;
+	std::vector<intptr_t> indicators_;
 };
