+++ cpp/cpp_odbc/Library/src/multi_value_buffer.cpp
@@ -58,12 +58,12 @@ char const * multi_value_buffer::data_pointer() const
 	return data_.data();
 }
 
-int64_t * multi_value_buffer::indicator_pointer()
+intptr_t * multi_value_buffer::indicator_pointer()
 {
 	return indicators_.data();
 }
 
-int64_t const * multi_value_buffer::indicator_pointer() const
+intptr_t const * multi_value_buffer::indicator_pointer() const
 {
 	return indicators_.data();
 }
